backend/requirements.txt
absl-py==2.3.1
altair==6.0.0
annotated-doc==0.0.4     
annotated-types==0.7.0   
anyio==4.11.0
asttokens==3.0.0
astunparse==1.6.3        
attrs==25.4.0
bleach==6.2.0
blinker==1.9.0
build==1.3.0
CacheControl==0.14.3     
cachetools==6.2.2        
certifi==2025.8.3        
cffi==2.0.0
charset-normalizer==3.4.3
cleo==2.1.0
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
crashtest==0.4.1
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.16
decorator==5.2.1
deprecation==2.1.0
distlib==0.4.0
dnspython==2.8.0
dulwich==0.24.2
email-validator==2.3.0
executing==2.2.1
fastapi==0.124.4
fastjsonschema==2.21.2
filelock==3.19.1
findpython==0.7.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.12.0
gast==0.7.0
gitdb==4.0.12
GitPython==3.1.45
google-pasta==0.2.0
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
h5py==3.15.1
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
installer==0.7.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.3.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.6.3
jupyter_core==5.8.1
kaggle==1.7.4.5
keras==3.12.0
keyring==25.6.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
matplotlib-inline==0.1.7
mdurl==0.1.2
ml_dtypes==0.5.4
mmh3==5.2.0
more-itertools==10.8.0
msgpack==1.1.1
multidict==6.7.0
namex==0.1.0
narwhals==2.13.0
nest-asyncio==1.6.0
nltk==3.9.2
numpy==2.3.3
opt_einsum==3.4.0
optree==0.18.0
packaging==25.0
pandas==2.3.3
parso==0.8.5
pbs-installer==2025.9.18
pillow==11.3.0
pkginfo==1.12.1.2
platformdirs==4.4.0
poetry==2.2.1
poetry-core==2.2.1
postgrest==2.26.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.32.1
psutil==7.0.0
pure_eval==0.2.3
pyarrow==22.0.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyiceberg==0.10.0
PyJWT==2.10.1
pyparsing==3.2.5
pyproject_hooks==1.2.0
pyroaring==1.0.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
python-slugify==8.0.4
pytz==2025.2
pyzmq==27.0.2
RapidFuzz==3.14.1
realtime==2.26.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.2.0
rpds-py==0.30.0
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
stack-data==0.6.3
starlette==0.50.0
storage3==2.26.0
streamlit==1.52.1
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.26.0
supabase-auth==2.26.0
supabase-functions==2.26.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
tensorflow==2.20.0
termcolor==3.2.0
text-unidecode==1.3
textblob==0.19.0
threadpoolctl==3.6.0
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
trove-classifiers==2025.9.11.17
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
virtualenv==20.34.0
watchdog==6.0.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.4
wheel==0.45.1
wordcloud==1.9.4
wrapt==2.0.1
yarl==1.22.0
zstandard==0.25.0
